V3.10 全链路修复 + 全量测试回归
## 本轮（2026-07-17）: 零宕机审查 8/8 + fix-prompts 8 项

### 零宕机审查: 6 Fatal + 2 Serious
- F1 ¥NaN: official.js/order-history.js price null 安全
- F2 白屏: quiz-result 分母零 + schema 对齐
- F3 无音频: reader audio_url 赋值 + play() 调用
- F4 null.find(): 5 处 || [] 保护
- F5 无限转圈: error-view + cancel 按钮
- F6 阅读清零: endSession(sid,0,words,minutes)
- S1 支付参数空: DepositPayResponse.pay_params
- S2 回调丢失: amount 补字段 + 分转元 + mock 修复

### fix-prompts 4/8 完成
- T7 seeder 统一 DEFAULTS: 37 键循环初始化 + 死键清理
- T8 ruff 92→0: E702/F401/F841/E701 全清
- T6 premium-hero::before 伪元素删除
- T4 iconfont: 285 emoji→62 icon 类, 31/31 WXML 文件, 降级策略

### 测试修复 (git checkout 回滚恢复)
- deposit: AsyncMock + REFUND_PENDING + pay_params 测试
- borrow: test_get_child_borrows 元组解包
- advancement: correct_answer 断言修正

## 前次会话 (2026-07-15): 全链条修复

### 第三方终审 9 项
- P0: Router ORM 下沉 Service (2处)
- P0: 押金退款 REFUND_PENDING + audit_refund 端点
- P0: 运营主体 COMPANY_NAME env var + 办学资质展示
- P1: 押金校验统一、JSON 日志 trace_id、innerHTML XSS

### 施工指令 6/6
- P1: observation-report 可选链 5处
- P1: levels.html loading 锁
- P1: users.html 24 内联函数外迁
- P2: 5 页定时器 onUnload 清理

### 安全 7/7 + 性能 N+1 19 处 + 微信合规 3 批
- 安全: 越权端点封闭、手机号脱敏、DEBUG 支付绕过移除
- 性能: 19 处 joinedload/批量查询/GROUP BY 聚合
- 合规: 登录隐私勾选、iOS 支付文案、sitemap 精细化

### 微信全链路
- 支付/押金/退款: 签名验签 + 幂等 + with_for_update
- Mock 支付网关: MOCK_PAYMENT guard
- SMS: 阿里云/腾讯云 SDK 实现 + ImportError 兜底

+++ backend/domain/activity/schemas.py
@@ -2,6 +2,7 @@
 """活动域 Pydantic 模型"""
 
 from datetime import datetime
+from decimal import Decimal
 from pydantic import Field
 from backend.common.base_schema import BaseSchema
 
@@ -12,7 +13,7 @@ class ActivityResponse(BaseSchema):
     description: str | None = None
     type: int | None = None
     is_free: int | None = 1
-    price: float | None = None
+    price: Decimal | None = None
     cover: str | None = None
     start_time: datetime | None = None
     end_time: datetime | None = None
